write event data to `detector_1_events` instead of `detector_1` (#433)
## Summary of changes

I think in a traditional nexus file events are put in the
`detector_1_events` group - this changes the nexus-writer to do the
same. Also removes a stray string in favour of a constant. event mode
files normally have a `detector_1` group but for retrospectively
histogrammed data.

## Instruction for review/testing

No testing needed really - the CI should be enough.